refactor: extract shared streaming export pipeline
